requirements.txt
beautifulsoup4==4.13.5
chromadb==1.1.0
GitPython==3.1.45
icecream==2.1.8
langchain==0.3.27
langchain-community==0.3.29
openai==1.108.1
pandas==2.3.2
PyDriller==2.9
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
tree-sitter==0.21.3
